+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## (Unreleased)[Unreleased]
+
+### Added
+- `ConfigLoaderAggregate` that aggregates configurations from multiple loaders.
+
 ## 1.0.0 - 2017-03-04
 
 ### Added
@@ -9,5 +14,6 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - `Interop\Container\ContainerInterface` as return type declaration in DI container factories
+
 
 [Unreleased]: https://github.com/nikolaposa/rate-limit/compare/1.0.0...HEAD